refactor(app): extract recovery cluster from PipelineQueue
Move loadSnapshot and recoverOrphanedRecordings out of PipelineQueue.swift
into a same-module extension file PipelineQueue+Recovery.swift.

WHY: after the stage-cluster extraction the class body was 465 effective
lines, still over the 400 type_body_length warning cap. Moving the
snapshot-restore and orphaned-recording recovery methods into an extension
drops the body to 366 effective lines, so the type_body_length suppression on
PipelineQueue is removed.

Pure move, no behavior change: the method bodies are verbatim. Visibility was
widened from private to internal only for logDir, which loadSnapshot reads
across the file boundary to locate the snapshot file; every other member the
moved methods touch (jobs, naming, processedLedger, eventLog, saveSnapshot,
triggerProcessing, cleanupStalePending, pendingSpeakerNamingJobs) was already
internal. cleanupStalePending stays in the main file: it is a speaker-naming
forwarder and tests call it directly.

Suppressions: type_body_length is now dropped. file_length stays suppressed on
PipelineQueue.swift (691 lines) and PipelineQueue+Stages.swift (644 lines)
since the orchestrator and its stage cluster are inherently long;
PipelineQueue+Recovery.swift (159 lines) needs no suppression.
