+++ app/MeetingTranscriber/Sources/PipelineQueue.swift
@@ -18,13 +18,17 @@ struct DiarizationRun {
 
 @MainActor
 @Observable
-// swiftlint:disable:next attributes type_body_length
+// swiftlint:disable:next attributes
 class PipelineQueue {
-    /// Internal setter (not `private(set)`) because the stage methods in
-    /// PipelineQueue+Stages.swift mutate jobs in place
-    /// (transcriptPath/protocolPath/namingSlug) across the file boundary.
+    /// Internal setter (not `private(set)`) because the stage and recovery
+    /// extension methods in sibling files (PipelineQueue+Stages.swift,
+    /// PipelineQueue+Recovery.swift) mutate jobs across the file boundary:
+    /// in-place field updates, snapshot restore (whole-array assign), and
+    /// recovery appends.
     var jobs: [PipelineJob] = []
-    private let logDir: URL
+    /// Internal (not private) because `loadSnapshot` in PipelineQueue+Recovery.swift
+    /// reads it to locate the snapshot file.
+    let logDir: URL
 
     /// File-backed skip list of already-processed recordings, consulted by
     /// `recoverOrphanedRecordings` so completed jobs aren't re-queued as
@@ -560,8 +564,10 @@ class PipelineQueue {
     // MARK: - Processing
 
     /// Kick off processing if not already running and there are waiting jobs.
-    /// Internal (not private) because `processNext` in PipelineQueue+Stages.swift
-    /// re-triggers the queue after finishing a job.
+    /// Internal (not private) because sibling-file extension methods re-trigger
+    /// the queue: `processNext` (PipelineQueue+Stages.swift) after finishing a
+    /// job, and `loadSnapshot` / `recoverOrphanedRecordings`
+    /// (PipelineQueue+Recovery.swift) after restoring or recovering jobs.
     func triggerProcessing() {
         guard !isProcessing else { return }
         guard pendingJobs.first != nil else { return }
@@ -579,154 +585,6 @@ class PipelineQueue {
     /// with auto-names, transitions them to .done, and deletes sidecar files.
     func cleanupStalePending(maxAge: TimeInterval = 86400) {
         naming.cleanupStalePending(pendingJobs: pendingSpeakerNamingJobs, maxAge: maxAge)
-    }
-
-    // MARK: - Snapshot Recovery
-
-    /// Load pipeline queue from the JSON snapshot written by `saveSnapshot()`.
-    /// Resets in-progress jobs to `.waiting`, discards `.done` jobs, and drops
-    /// jobs whose `mixPath` no longer exists on disk.
-    func loadSnapshot() {
-        var loaded: [PipelineJob]
-        do {
-            guard let decoded = try PipelineSnapshot.load(from: logDir) else {
-                logger.info("No pipeline snapshot to restore")
-                return
-            }
-            loaded = decoded
-        } catch {
-            logger.error("Failed to load pipeline snapshot: \(error.localizedDescription, privacy: .public)")
-            return
-        }
-
-        // Reset active states back to waiting
-        for i in loaded.indices {
-            switch loaded[i].state {
-            case .transcribing, .diarizing, .generatingProtocol:
-                loaded[i].state = .waiting
-
-            default:
-                break
-            }
-        }
-
-        // Discard done jobs
-        loaded.removeAll { $0.state == .done }
-
-        // Discard jobs whose audio file no longer exists, EXCEPT
-        // .speakerNamingPending — those have their own slug-based
-        // `_16k.wav` sidecar and don't need the original mix.wav.
-        // Paired imports with nil mixPath: keep them — `appPath` is the
-        // ground-truth source and it's checked at processNext time.
-        loaded.removeAll { job in
-            guard let mixPath = job.mixPath else { return false }
-            return job.state != .speakerNamingPending
-                && !FileManager.default.fileExists(atPath: mixPath.path)
-        }
-
-        guard !loaded.isEmpty else {
-            logger.info("Snapshot loaded but no recoverable jobs")
-            return
-        }
-
-        jobs = loaded
-
-        // Rebuild the session's naming cache from disk for
-        // .speakerNamingPending jobs.
-        for job in jobs where job.state == .speakerNamingPending {
-            if let slug = job.namingSlug, naming.restore(jobID: job.id, slug: slug) {
-                continue
-            }
-            // Naming data lost — transition to done
-            logger.warning("Naming data not found for job \(job.id), marking as done")
-            if let idx = jobs.firstIndex(where: { $0.id == job.id }) {
-                jobs[idx].state = .done
-            }
-        }
-
-        saveSnapshot()
-        cleanupStalePending()
-        logger.info("Restored \(loaded.count) jobs from snapshot")
-        triggerProcessing()
-        // Auto-popup the naming dialog if any restored job is still
-        // waiting for confirmation. Same notification as the in-pipeline
-        // pop, so MeetingTranscriberApp brings the window forward.
-        if !pendingSpeakerNamingJobs.isEmpty {
-            NotificationCenter.default.post(name: .showSpeakerNaming, object: nil)
-        }
-    }
-
-    // MARK: - Orphaned Recording Recovery
-
-    /// Scan `recordingsDir` for `*_mix.wav` files not tracked by any loaded job.
-    /// Creates recovery jobs for untracked recordings younger than `maxAge`.
-    /// Skips files that were already successfully processed (tracked in processed_recordings.json).
-    ///
-    /// The directory scan + per-file `attributesOfItem` calls run on a
-    /// detached task — startup callers don't block the UI on a potentially
-    /// slow filesystem (e.g. iCloud-backed recordings dir). Mutations to
-    /// `jobs` and the snapshot still happen on the main actor.
-    func recoverOrphanedRecordings(
-        recordingsDir: URL = AppPaths.recordingsDir,
-        maxAge: TimeInterval = 86400,
-    ) async {
-        // One-time migration: seed processed list with existing recordings
-        // Only for the default recordings directory (not test overrides)
-        if recordingsDir == AppPaths.recordingsDir {
-            await processedLedger.migrate(recordingsDir: recordingsDir)
-        }
-
-        let trackedPaths = Set(jobs.compactMap { $0.mixPath?.standardizedFileURL.path })
-        let ledger = processedLedger
-
-        // Off-main: directory scan + processed-list read + per-file
-        // attributesOfItem probes + filtering all happen here.
-        let candidates: [PairedRecordingResolver.Group] = await Task.detached(priority: .utility) {
-            let fm = FileManager.default
-            guard let entries = try? fm.contentsOfDirectory(
-                at: recordingsDir,
-                includingPropertiesForKeys: [.fileSizeKey, .creationDateKey],
-            ) else { return [] }
-            let processedPaths = ledger.load()
-            let now = Date()
-            return PairedRecordingResolver.resolve(urls: entries).paired.filter { group in
-                // Groups without a real mix file (app+mic-only paired imports)
-                // aren't recoverable from the dir scan alone.
-                guard let mixURL = group.mix else { return false }
-                let stdPath = mixURL.standardizedFileURL.path
-                guard !trackedPaths.contains(stdPath) else { return false }
-                guard !processedPaths.contains(stdPath) else { return false }
-                let attrs = try? fm.attributesOfItem(atPath: mixURL.path)
-                if let created = attrs?[.creationDate] as? Date,
-                   now.timeIntervalSince(created) > maxAge {
-                    return false
-                }
-                // Header-only WAVs are 44 bytes.
-                if let size = attrs?[.size] as? Int, size <= 44 {
-                    return false
-                }
-                return true
-            }
-        }.value
-
-        guard !candidates.isEmpty else { return }
-
-        for group in candidates {
-            guard let mixURL = group.mix else { continue }
-            let job = PipelineJob(
-                meetingTitle: "Recovered Recording (\(group.stem))",
-                appName: "Unknown",
-                mixPath: mixURL,
-                appPath: group.app,
-                micPath: group.mic,
-                micDelay: 0,
-            )
-            jobs.append(job)
-            eventLog.append(jobID: job.id, event: "recovered", from: nil, to: .waiting)
-        }
-        saveSnapshot()
-        logger.info("Recovered \(candidates.count) orphaned recording(s)")
-        triggerProcessing()
     }
 
     // MARK: - Log Directory
